🎉 FEAT(테스트코드): order
order 테스트 코드 추가,
쿠폰 상태 변경 메서드 추가,
카카오페이 환불 예외처리 추가

+++ src/main/java/com/supernova/fashionnova/domain/order/dto/AllOrderResponseDto.java
@@ -13,9 +13,9 @@ import lombok.NoArgsConstructor;
 public class AllOrderResponseDto {
 
   private Long orderId;
-  private Long totalPrice;
-  private Long cost;
-  private int discount;
+  private double totalPrice;
+  private double cost;
+  private double discount;
   private Long usedMileage;
   private String address;
   private LocalDateTime createdAt;
